refactor: move role to package
